Move node implementation to custom plugin
