+++ conftest.py
@@ -1,0 +1,6 @@
+"""This file sets up custom plugins.
+
+https://docs.pytest.org/en/stable/writing_plugins.html
+
+"""
+pytest_plugins = "node_plugin"
